Spot miner experiment

+++ modules/setup-miner.sh
@@ -3,18 +3,25 @@
 export PS4='setup-miner.sh: '
 set -ex
 
-for snap in lxd core20 snapd; do
-  snap remove $snap
-done
-systemctl stop snapd.{service,socket} snapd.{seeded,apparmor}.service
-apt-get autoremove -y --purge snapd
-rm -rf /root/snap
+if [ -f /etc/amazon-linux-release ]; then
+  # Amazon Linux 2023
+  dnf install docker
+  systemctl enable docker
+else
+  # Ubuntu 22.04
+  for snap in lxd core20 snapd; do
+    snap remove $snap
+  done
+  systemctl stop snapd.{service,socket} snapd.{seeded,apparmor}.service
+  apt-get autoremove -y --purge snapd
+  rm -rf /root/snap
 
-apt-get autoremove -y --purge modemmanager
+  apt-get autoremove -y --purge modemmanager
 
-apt-get update
-#apt-get upgrade -y
-apt-get install -y docker.io
+  apt-get update
+  #apt-get upgrade -y
+  apt-get install -y docker.io
+fi
 
 service=xmrig
 secret_name=config_passphrase
